# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Libre-Q API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/admin/studies": {
      "post": {
        "tags": [
          "admin-studies"
        ],
        "summary": "Create Study",
        "description": "Create a new study in the active project.\n\nRequires Owner or Researcher project role.",
        "operationId": "create_study_api_admin_studies_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "X-Project-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Project-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudyCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudyRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "admin-studies"
        ],
        "summary": "List Studies",
        "description": "List studies in the active project with pagination.",
        "operationId": "list_studies_api_admin_studies_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "X-Project-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Project-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_StudyRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/studies/{slug}": {
      "get": {
        "tags": [
          "admin-studies"
        ],
        "summary": "Get Study",
        "description": "Get study details.",
        "operationId": "get_study_api_admin_studies__slug__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The slug of the study",
              "title": "Slug"
            },
            "description": "The slug of the study"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudyRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "admin-studies"
        ],
        "summary": "Update Study",
        "description": "Update study configuration (draft only).",
        "operationId": "update_study_api_admin_studies__slug__patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The slug of the study",
              "title": "Slug"
            },
            "description": "The slug of the study"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudyUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudyRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "admin-studies"
        ],
        "summary": "Delete Study",
        "description": "Delete a study (Superuser only, and must be Archived).",
        "operationId": "delete_study_api_admin_studies__slug__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The slug of the study",
              "title": "Slug"
            },
            "description": "The slug of the study"
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/studies/{slug}/validate": {
      "post": {
        "tags": [
          "admin-studies"
        ],
        "summary": "Validate Study",
        "description": "Check if study is ready for activation.",
        "operationId": "validate_study_api_admin_studies__slug__validate_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The slug of the study",
              "title": "Slug"
            },
            "description": "The slug of the study"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  },
                  "title": "Response Validate Study Api Admin Studies  Slug  Validate Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/studies/{slug}/state": {
      "post": {
        "tags": [
          "admin-studies"
        ],
        "summary": "Change Study State",
        "description": "Change study state (Draft <-> Active <-> Closed <-> Archived).",
        "operationId": "change_study_state_api_admin_studies__slug__state_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The slug of the study",
              "title": "Slug"
            },
            "description": "The slug of the study"
          },
          {
            "name": "new_state",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/StudyState"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudyRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/studies/{slug}/reset": {
      "post": {
        "tags": [
          "admin-studies"
        ],
        "summary": "Reset Study Participants",
        "description": "Delete all participants for the study (Owner only).",
        "operationId": "reset_study_participants_api_admin_studies__slug__reset_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The slug of the study",
              "title": "Slug"
            },
            "description": "The slug of the study"
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/studies/{slug}/import-concourse": {
      "post": {
        "tags": [
          "admin-studies"
        ],
        "summary": "Import From Concourse",
        "description": "Import concourse items into a study as statements (copies, no reference).",
        "operationId": "import_from_concourse_api_admin_studies__slug__import_concourse_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The slug of the study",
              "title": "Slug"
            },
            "description": "The slug of the study"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConcourseImportToStudy"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudyRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/studies/{slug}/stale-statements": {
      "get": {
        "tags": [
          "admin-studies"
        ],
        "summary": "Check Stale Statements",
        "description": "Check which imported statements have stale concourse sources.",
        "operationId": "check_stale_statements_api_admin_studies__slug__stale_statements_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The slug of the study",
              "title": "Slug"
            },
            "description": "The slug of the study"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StaleStatementRead"
                  },
                  "title": "Response Check Stale Statements Api Admin Studies  Slug  Stale Statements Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/studies/{slug}/sync-statement/{statement_id}": {
      "post": {
        "tags": [
          "admin-studies"
        ],
        "summary": "Sync Statement From Concourse",
        "description": "Sync a single statement's text from its concourse source.",
        "operationId": "sync_statement_from_concourse_api_admin_studies__slug__sync_statement__statement_id__post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The slug of the study",
              "title": "Slug"
            },
            "description": "The slug of the study"
          },
          {
            "name": "statement_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Statement Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudyRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/studies/{slug}/sync-all-stale": {
      "post": {
        "tags": [
          "admin-studies"
        ],
        "summary": "Sync All Stale Statements",
        "description": "Sync all stale statements from their concourse sources.",
        "operationId": "sync_all_stale_statements_api_admin_studies__slug__sync_all_stale_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The slug of the study",
              "title": "Slug"
            },
            "description": "The slug of the study"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudyRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConcourseImportToStudy": {
        "properties": {
          "concourse_id": {
            "type": "integer",
            "title": "Concourse Id"
          },
          "item_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "maxItems": 500,
            "minItems": 1,
            "title": "Item Ids"
          },
          "code_prefix": {
            "type": "string",
            "maxLength": 20,
            "title": "Code Prefix",
            "default": ""
          },
          "replace_existing": {
            "type": "boolean",
            "title": "Replace Existing",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "concourse_id",
          "item_ids"
        ],
        "title": "ConcourseImportToStudy",
        "description": "Schema for importing concourse items into a study as statements."
      },
...
      "GridColumn": {
        "properties": {
          "score": {
            "type": "integer",
            "title": "Score"
          },
          "capacity": {
            "type": "integer",
            "title": "Capacity"
          }
        },
        "type": "object",
        "required": [
          "score",
          "capacity"
        ],
        "title": "GridColumn",
        "description": "Schema defining a column in the sorting grid."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PaginatedResponse_StudyRead_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/StudyRead"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "PaginatedResponse[StudyRead]"
      },
...
      "StudyCreate": {
        "properties": {
          "slug": {
            "type": "string",
            "maxLength": 100,
            "minLength": 3,
            "pattern": "^[a-z0-9-]+$",
            "title": "Slug"
          },
          "state": {
            "$ref": "#/components/schemas/StudyState",
            "default": "draft"
          },
          "grid_config": {
            "items": {
              "$ref": "#/components/schemas/GridColumn"
            },
            "type": "array",
            "title": "Grid Config"
          },
          "presort_config": {
            "additionalProperties": true,
            "type": "object",
            "title": "Presort Config"
          },
          "postsort_config": {
            "additionalProperties": true,
            "type": "object",
            "title": "Postsort Config"
          },
          "branding": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BrandingBase"
              },
              {
                "type": "null"
              }
            ]
          },
          "default_language": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 5
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Language"
          },
          "show_statement_codes": {
            "type": "boolean",
            "title": "Show Statement Codes",
            "default": false
          },
          "randomize_statement_order": {
            "type": "boolean",
            "title": "Randomize Statement Order",
            "default": false
          },
          "symmetry_lock": {
            "type": "boolean",
            "title": "Symmetry Lock",
            "default": true
          },
          "start_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Date"
          },
          "end_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          },
          "translations": {
            "items": {
              "$ref": "#/components/schemas/StudyTranslationCreate"
            },
            "type": "array",
            "title": "Translations"
          },
          "statements": {
            "items": {
              "$ref": "#/components/schemas/StatementCreate"
            },
            "type": "array",
            "title": "Statements",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "slug",
          "grid_config",
          "presort_config",
          "postsort_config",
          "translations"
        ],
        "title": "StudyCreate",
        "description": "Schema for creating a study."
      },
...
      "StudyState": {
        "type": "string",
        "enum": [
          "draft",
          "active",
          "paused",
          "closed",
          "archived"
        ],
        "title": "StudyState",
        "description": "Enum for study lifecycle states."
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "api/token"
          }
        }
      }
    }
  }
}